Move PlayerCount to playersRule from player
