+++ src/domain/models/player/index.ts
@@ -1,6 +1,5 @@
 export * from './IPlayerFactory';
 export * from './IPlayerRepository';
 export { default as Player } from './Player';
-export { default as PlayerCount } from './PlayerCount';
 export { default as PlayerId } from './PlayerId';
 export { default as PlayerName } from './PlayerName';
