+++ src/domain/models/playersRule/PlayerCount.ts
