+++ src/domain/models/playersRule/index.ts
@@ -1,5 +1,6 @@
 export * from './IPlayerRuleFactory';
 export * from './IPlayerRuleRepository';
+export { default as PlayerCount } from './PlayerCount';
 export { default as PlayersRule } from './PlayersRule';
 export { default as PlayersRuleId } from './PlayersRuleId';
 export { default as PlayersRuleType } from './PlayersRuleType';
